backend/requirements.txt
fastapi
uvicorn
openai
langchain
langchain-community
langchain-chroma
langchain-openai
chromadb
pydantic
python-dotenv
tiktoken
beautifulsoup4
